% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentreFreqMhzRange.r
\name{centre_freq_mhz_range}
\alias{centre_freq_mhz_range}
\title{Author: Sarkis Jacob
CRC/OTTAWA - Canada}
\usage{
centre_freq_mhz_range(centre_freq_mhz_i, centre_freq_mhz_f)
}
\arguments{
\item{centre_freq_mhz_i}{Defines the initial value of the centre frequency range.}

\item{centre_freq_mhz_f}{Defines the final value of the centre frequency range.}
}
\value{
Spark dataframe
}
\description{
January 29, 2019 --
This program filters dataset within a range of frequencies defined by the user.
}
\examples{
centre_freq_mhz_range(centre_freq_mhz_i,centre_freq_mhz_f)
}
